pyproject.toml
[tool.poetry]
name = "metastatic_models"
version = "0.1.0"
description = "model to predict probability of liver metastasis using mutation information"
authors = ["Amir Fatemi <amir2520@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
imbalanced-learn = "0.9.1"
imblearn = "0.0"
joblib = "1.2.0"
pandas = "1.4.2"
scikit-learn = "1.2.0"
setuptools = "61.2.0"
numpy = ">=1.24,<1.26"
mlflow = "~=2.5.0"
notebook = "~=6.5"
psycopg2-binary = "~=2.9"
pydantic = "~=1.10"
hydra-core = "~=1.3"
fsspec = {version = "~=2023.6", extras = ['gcs']}
gcsfs = "~=2023.6"
fastapi = "~=0.104"
uvicorn = {version = "0.30.6", extras = ["standard"]}  

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
